refactor(aviation): move up_command + formatters out of mod.rs
