+++ src/aviation/commands/mod.rs
@@ -3,3 +3,4 @@ pub mod flights_above;
 pub mod random_flight;
 pub mod track;
 pub mod untrack;
+pub mod up;
